# tides-server/swagger/swagger.yml
swagger: "2.0"
info:
  title: 'CloudTides-Server'
  version: '0.0.1'
basePath: /api/v1
paths:
...
  /application/instance/file/{uid}/{token}/{name}:
    get:
      operationId: 'downInstanceFile'
      description: 'down files to instances'
      tags:
        - 'application'
      produces:
        - 'application/octet-stream'
      parameters:
        - type: string
          in: path
          name: token
          required: true
        - type: string
          in: path
          name: name
          required: true
        - type: string
          in: path
          name: uid
          required: true
      responses:
        200:
          description: 'OK'
  /application/instance/file/{token}:
    get:
      operationId: 'listInstanceFiles'
      description: 'list files to instances'
      tags:
        - 'application'
      produces:
        - 'application/json'
      parameters:
        - type: string
          in: path
          name: token
          required: true
      responses:
        200:
          description: 'OK'
          schema:
            type: array
            items:
              type: object
              properties:
                filename:
                  type: string
                filesize:
                  type: string
                createtime:
                  type: string
                downlink:
                  type: string

    post:
      operationId: 'uploadInstanceFile'
      description: 'upload file to instance'
      tags:
        - 'application'
      produces:
        - 'application/json'
      consumes:
        - 'multipart/form-data'
      parameters:
        - name: file
          in: formData
          type: file
          required: true
        - in: path
          name: token
          type: string
          required: true

      responses:
        200:
          description: 'OK'
          schema:
            type: object
            properties:
              downlink:
                type: string
...
  /application/instance:
    get:
      operationId: 'listApplicationInstance'
      description: 'list application instance'
      tags:
        - 'application'
      produces:
        - 'application/json'
      responses:
        200:
          description: 'OK'
          schema:
            type: array
            items:
              type: object
              properties:
                instanceName:
                  type: string
                link:
                  type: string
                token:
                  type: string
                appType:
                  type: string
                port:
                  type: string
                sshHost:
                  type: string
                sshUser:
                  type: string
                sshPassword:
                  type: string
                sshPort:
                  type: string
                createAt:
                  type: string
                runningTime:
                  type: string
                extra:
                  type: object
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
    post:
      operationId: 'createApplicationInstance'
      description: 'create application instance'
      tags:
        - 'application'
      produces:
        - 'application/json'
      parameters:
        - in: body
          name: reqBody
          schema:
            type: object
            properties:
              instanceName:
                type: string
              port:
                type: string
              sshHost:
                type: string
              sshUser:
                type: string
              sshPassword:
                type: string
              sshPort:
                type: integer
              appType:
                type: string
      responses:
        200:
          description: 'OK'
          schema:
            type: object
            properties:
              link:
                type: string
              token:
                type: string
              extra:
                type: object
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
          schema:
            type: object
            properties:
              msg:
                type: string
    put:
      operationId: 'updateApplicationInstance'
      description: 'update application instance'
      tags:
        - 'application'
      produces:
        - 'application/json'
      parameters:
        - in: body
          name: reqBody
          schema:
            type: object
            properties:
              instanceName:
                type: string
              port:
                type: string
              sshHost:
                type: string
              sshUser:
                type: string
              sshPassword:
                type: string
              sshPort:
                type: integer
              appType:
                type: string
              token:
                type: string
      responses:
        200:
          description: 'OK'
          schema:
            type: object
            properties:
              link:
                type: string
              token:
                type: string
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
          schema:
            type: object
            properties:
              msg:
                type: string
  /ws/application/instance/{token}:
    get:
      operationId: 'WsWatchApplicationInstanceLogs'
      description: 'Websocket Watch application instance logs'
      tags:
        - 'application'
      produces:
        - 'application/json'
      parameters:
        - type: string
          in: path
          name: token
          required: true
      responses:
        200:
          description: 'OK'
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
  /application/instance/{token}:
    get:
      operationId: 'watchApplicationInstanceLogs'
      description: 'watch application instance logs'
      tags:
        - 'application'
      produces:
        - 'application/json'
      parameters:
        - type: string
          in: path
          name: token
          required: true
      responses:
        200:
          description: 'OK'
          schema:
            type: array
            items:
              type: object
              properties:
                leve:
                  type: string
                date:
                  type: string
                time:
                  type: string
                source:
                  type: string
                content:
                  type: string
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
    delete:
      operationId: 'deleteApplicationInstance'
      description: 'delete application instance'
      tags:
        - 'application'
      produces:
        - 'application/json'
      parameters:
        - type: string
          in: path
          name: token
          required: true
      responses:
        200:
          description: 'OK'
          schema:
            type: object
            properties:
              msg:
                type: string
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
  /application/instance/hosts:
    get:
      operationId: 'achieveHost'
      description: 'achieve host'
      tags:
        - 'application'
      produces:
        - 'application/json'
      responses:
        200:
          description: 'OK'
          schema:
            type: array
            items:
              type: object
              properties:
                address:
                  type: string
                sshUser:
                  type: string
                sshPort:
                  type: string
                sshPass:
                  type: string
        401:
          description: 'Unauthorized'
        403:
          description: 'Forbidden'
...